+++ src/StakingContract.sol
@@ -2,19 +2,19 @@
 pragma solidity ^0.8.13;
 
 interface IMToken {
-    function mint(address to, uint amount) external;
+    function mint(address to, uint256 amount) external;
 }
 
 contract StakingContract {
     IMToken public mToken;
-    uint public totalStaked;
-    mapping(address => uint) public stakedBalances;
-    mapping(address => uint) public unclaimedRewards;
-    mapping(address => uint) public lastUpdatedTime;
+    uint256 public totalStaked;
+    mapping(address => uint256) public stakedBalances;
+    mapping(address => uint256) public unclaimedRewards;
+    mapping(address => uint256) public lastUpdatedTime;
 
-    struct WithdrawlRequest{
-        uint amount;
-        uint releaseTime;
+    struct WithdrawlRequest {
+        uint256 amount;
+        uint256 releaseTime;
     }
 
     mapping(address => WithdrawlRequest[]) public pendingWithdrawls;
@@ -28,10 +28,8 @@ contract StakingContract {
         // Calculate rewards
         if (lastUpdatedTime[msg.sender] == 0) {
             lastUpdatedTime[msg.sender] = block.timestamp;
-        }
-
-        else {
-            uint reward = getRewards(msg.sender);
+        } else {
+            uint256 reward = getRewards(msg.sender);
             unclaimedRewards[msg.sender] = reward;
             lastUpdatedTime[msg.sender] = block.timestamp;
         }
@@ -39,37 +37,36 @@ contract StakingContract {
         totalStaked += msg.value;
     }
 
-    function unstake(uint _amount) public {
+    function unstake(uint256 _amount) public {
         require(stakedBalances[msg.sender] >= _amount, "insufficient staked balance");
 
         //Calculate rewards
-        uint reward = getRewards(msg.sender);
+        uint256 reward = getRewards(msg.sender);
         unclaimedRewards[msg.sender] = reward;
         lastUpdatedTime[msg.sender] = block.timestamp;
         stakedBalances[msg.sender] -= _amount;
         totalStaked -= _amount;
         pendingWithdrawls[msg.sender].push(WithdrawlRequest(_amount, block.timestamp + 1 minutes));
-        
     }
 
-    function getRewards(address _address) public view returns(uint) {
-        uint currentReward = unclaimedRewards[_address];
-        uint timePassed = (block.timestamp - lastUpdatedTime[_address])/60;
-        uint newReward = stakedBalances[_address] * timePassed;
+    function getRewards(address _address) public view returns (uint256) {
+        uint256 currentReward = unclaimedRewards[_address];
+        uint256 timePassed = (block.timestamp - lastUpdatedTime[_address]) / 60;
+        uint256 newReward = stakedBalances[_address] * timePassed;
         return currentReward + newReward;
     }
 
     function claimRewards() public {
-        uint reward = getRewards(msg.sender);
+        uint256 reward = getRewards(msg.sender);
         unclaimedRewards[msg.sender] = 0;
         lastUpdatedTime[msg.sender] = block.timestamp;
         mToken.mint(msg.sender, reward);
     }
 
-    function withdraw() public{
-        uint totalWithdrawlAmount;
+    function withdraw() public {
+        uint256 totalWithdrawlAmount;
         WithdrawlRequest[] memory reqs = pendingWithdrawls[msg.sender];
-        for (uint i = 0; i < reqs.length; i++) {
+        for (uint256 i = 0; i < reqs.length; i++) {
             if (block.timestamp >= reqs[i].releaseTime) {
                 totalWithdrawlAmount += reqs[i].amount;
                 delete pendingWithdrawls[msg.sender][i];
